+++ dsh-api/src/dsh_api_tenant.rs
@@ -1,6 +1,6 @@
 //! # Client tenant
 use crate::platform::DshPlatform;
-use crate::{guid_environment_variable, DshApiError, ENV_VAR_TENANT};
+use crate::{DshApiError, ENV_VAR_TENANT};
 use lazy_static::lazy_static;
 use log::info;
 use std::env;
@@ -9,7 +9,6 @@ use std::fmt::{Display, Formatter};
 #[derive(Clone, Debug)]
 pub struct DshApiTenant {
   name: String,
-  guid: u16,
   platform: DshPlatform,
 }
 
@@ -18,7 +17,6 @@ impl DshApiTenant {
   ///
   /// # Parameters
   /// * `name` - client tenant's name
-  /// * `guid` - client tenant's group and user id (must be the same on DSH)
   /// * `platform` - target platform for the API
   ///
   /// # Examples
@@ -28,9 +26,8 @@ impl DshApiTenant {
   /// # use dsh_api::dsh_api_tenant::DshApiTenant;
   /// # use dsh_api::platform::DshPlatform;
   /// let name = String::from("my-tenant");
-  /// let guid = 1234;
   /// let platform = DshPlatform::try_from("nplz")?;
-  /// let dsh_api_tenant = DshApiTenant::new(name, guid, platform);
+  /// let dsh_api_tenant = DshApiTenant::new(name, platform);
   /// assert_eq!(
   ///   dsh_api_tenant.platform().internal_service_domain("my-service"),
   ///   "my-service.marathon.mesos".to_string()
@@ -38,16 +35,14 @@ impl DshApiTenant {
   /// # Ok(())
   /// # }
   /// ```
-  pub fn new(name: String, guid: u16, platform: DshPlatform) -> Self {
-    Self { name, guid, platform }
+  pub fn new(name: String, platform: DshPlatform) -> Self {
+    Self { name, platform }
   }
 
   /// # Create new DSH API tenant from tenant's name
   ///
   /// This factory function will attempt to create a `DshapiTenant` instance from the provided
   /// tenant's name.
-  /// The tenant's group and user ids will be read from the
-  /// environment variable `DSH_API_GUID_[TENANT]`.
   /// The platform will be read from the
   /// environment variable `DSH_API_PLATFORM`.
   /// The function will return an `Error<String>` if the environment variables are not set
@@ -74,16 +69,14 @@ impl DshApiTenant {
   /// # }
   /// ```
   pub fn from_tenant(tenant_name: String) -> Result<Self, DshApiError> {
-    let guid = guid_from_tenant_name(tenant_name.as_str())?;
     let platform = DshPlatform::default();
-    Ok(DshApiTenant::new(tenant_name, guid, platform))
+    Ok(DshApiTenant::new(tenant_name, platform))
   }
 
   /// # Create new DSH API tenant from tenant's name and platform
   ///
   /// This factory function will attempt to create a `DshapiTenant` instance from the provided
-  /// tenant's name. The group and user ids will be read from the
-  /// environment variable `DSH_API_GUID_[TENANT]`.
+  /// tenant's name.
   /// The function will return an `Error<String>` if the environment variable is not set.
   ///
   /// # Parameters
@@ -109,16 +102,13 @@ impl DshApiTenant {
   /// # }
   /// ```
   pub fn from_tenant_and_platform(tenant_name: String, platform: DshPlatform) -> Result<Self, DshApiError> {
-    let guid = guid_from_tenant_name(tenant_name.as_str())?;
-    Ok(DshApiTenant::new(tenant_name, guid, platform))
+    Ok(DshApiTenant::new(tenant_name, platform))
   }
 
   /// # Create new DSH API tenant from platform
   ///
   /// This factory function will attempt to create a `DshapiTenant` instance from the provided
-  /// `platform`. The tenant's name and group and user ids will be read from the
-  /// environment variables `DSH_API_TENANT` and
-  /// `DSH_API_GUID_[TENANT]`.
+  /// `platform`. The tenant's name will be read from the environment variable.
   /// The function will return an `Error<String>` if the environment variables are not set.
   ///
   /// # Parameters
@@ -147,13 +137,12 @@ impl DshApiTenant {
       Ok(name) => name,
       Err(_) => return Err(DshApiError::Configuration(format!("environment variable {} not set", ENV_VAR_TENANT))),
     };
-    let guid = guid_from_tenant_name(tenant_name.as_str())?;
-    Ok(DshApiTenant::new(tenant_name, guid, platform))
+    Ok(DshApiTenant::new(tenant_name, platform))
   }
 
   /// Returns the default tenant
   ///
-  /// This method will read the tenant name, guid and platform form the respective
+  /// This method will read the tenant name and platform form the respective
   /// environment variables and will create a`DshApiTenant` if possible. It will return an
   /// `Error<String>` when one or more of these the environment variables is not set or
   /// contains an undefined value.
@@ -164,9 +153,8 @@ impl DshApiTenant {
   ///   contains an undefined value
   pub fn try_default() -> Result<Self, String> {
     let tenant_name = get_default_tenant_name()?;
-    let guid = guid_from_tenant_name(tenant_name.as_str())?;
     let platform = DshPlatform::try_default()?;
-    Ok(DshApiTenant::new(tenant_name, guid, platform))
+    Ok(DshApiTenant::new(tenant_name, platform))
   }
 
   /// Returns the client's platform
@@ -178,17 +166,12 @@ impl DshApiTenant {
   pub fn name(&self) -> &String {
     &self.name
   }
-
-  /// Returns the client's guid
-  pub fn guid(&self) -> u16 {
-    self.guid
-  }
 }
 
 impl Default for DshApiTenant {
   /// Returns the default tenant
   ///
-  /// This method will read the tenant name, guid and platform form the respective
+  /// This method will read the tenant name and platform form the respective
   /// environment variables and will create a`DshApiTenant` if possible.
   ///
   /// # Panics
@@ -207,7 +190,7 @@ impl Default for DshApiTenant {
 
 impl Display for DshApiTenant {
   fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-    write!(f, "{}:{}@{}", self.name, self.guid, self.platform)
+    write!(f, "{}@{}", self.name, self.platform)
   }
 }
 
@@ -215,30 +198,6 @@ lazy_static! {
   pub static ref DEFAULT_DSH_API_TENANT: DshApiTenant = DshApiTenant::default();
 }
 
-fn parse_and_validate_guid(guid: String) -> Result<u16, DshApiError> {
-  match guid.parse::<u16>() {
-    Ok(guid) => {
-      if guid > 0 && guid < 60000 {
-        Ok(guid)
-      } else {
-        Err(DshApiError::Configuration(format!("guid {} not in range (1 <= guid < 60000)", guid)))
-      }
-    }
-    Err(_) => Err(DshApiError::Configuration(format!("could not parse guid '{}'", guid))),
-  }
-}
-
 fn get_default_tenant_name() -> Result<String, DshApiError> {
   env::var(ENV_VAR_TENANT).map_err(|_| DshApiError::Configuration(format!("environment variable {} not set", ENV_VAR_TENANT)))
-}
-
-fn guid_from_tenant_name(tenant_name: &str) -> Result<u16, DshApiError> {
-  let guid_env = guid_environment_variable(tenant_name);
-  match env::var(&guid_env) {
-    Ok(guid) => match parse_and_validate_guid(guid) {
-      Ok(guid) => Ok(guid),
-      Err(error) => Err(DshApiError::Configuration(format!("{} in environment variable {}", error, guid_env))),
-    },
-    Err(_) => Err(DshApiError::Configuration(format!("environment variable {} not set", guid_env))),
-  }
 }