+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to the DSH Api Client project will be documented in this fil
 * All API methods are now generated from the openapi specification.
   Hence, many methods now have a different name.
 * Renamed `api_version()` function to `openapi_version()`.
+* Removed group and user id from tenant data.
 
 ### Added